test(plugins): exercise profiling through the native module loader
