+++ src/plugins/plugin-module-loader-cache.test.ts
@@ -988,87 +988,6 @@ describe("getCachedPluginModuleLoader", () => {
     ]);
   });
 
-  it("reuses successful source-transform module exports inside one loader", async () => {
-    vi.stubEnv("OPENCLAW_DIAGNOSTICS", "plugin.load-profile");
-    const output = vi.spyOn(console, "error").mockImplementation(() => {});
-    const moduleExport = { marker: "source-cached" };
-    const fromSourceTransformer = vi.fn(() => moduleExport);
-    const createJiti = vi.fn(() => fromSourceTransformer);
-    const nativeStub = vi.fn(() => ({ ok: true, moduleExport: { fromNative: true } }));
-    vi.doMock("./native-module-require.js", async (importOriginal) => ({
-      ...(await importOriginal<typeof import("./native-module-require.js")>()),
-      tryNativeRequireJavaScriptModule: nativeStub,
-    }));
-    const { getCachedPluginModuleLoader, getPluginModuleLoaderStats } =
-      await importPluginModuleLoader("./plugin-module-loader-cache.js?scope=source-export-cache");
-
-    const cache = new Map();
-    const loader = getCachedPluginModuleLoader({
-      cache,
-      modulePath: "/repo/extensions/demo/api.ts",
-      importerUrl: "file:///repo/src/plugins/bundled-capability-runtime.ts",
-      loaderFilename: "file:///repo/src/plugins/bundled-capability-runtime.ts",
-      tryNative: false,
-      createLoader: asPluginModuleLoaderFactory(createJiti),
-    });
-
-    expect(loader("/repo/extensions/demo/api.ts")).toBe(moduleExport);
-    expect(loader("/repo/extensions/demo/api.ts")).toBe(moduleExport);
-    expect(loader("/repo/extensions/demo/other.ts")).toBe(moduleExport);
-    expect(nativeStub).not.toHaveBeenCalled();
-    expect(createJiti).toHaveBeenCalledOnce();
-    expect(fromSourceTransformer).toHaveBeenCalledTimes(2);
-    const stats = expectStats(getPluginModuleLoaderStats(), {
-      calls: 2,
-      nativeHits: 0,
-      nativeMisses: 0,
-      sourceTransformFallbacks: 0,
-      sourceTransformForced: 2,
-    });
-    expect(stats.topSourceTransformTargets).toEqual([
-      { target: "/repo/extensions/demo/api.ts", count: 1 },
-      { target: "/repo/extensions/demo/other.ts", count: 1 },
-    ]);
-    expect(output).toHaveBeenCalledExactlyOnceWith(
-      expect.stringMatching(
-        /^\[plugin-load-profile\] phase=source-transform-prepare plugin=\(core\) elapsedMs=\d+\.\d source=\(module\)$/,
-      ),
-    );
-  });
-
-  it("preserves transform preparation errors while profiling", async () => {
-    vi.stubEnv("OPENCLAW_DIAGNOSTICS", "plugin.load-profile");
-    const output = vi.spyOn(console, "error").mockImplementation(() => {});
-    const expectedError = new Error("fixture preparation failed");
-    const createLoader = vi.fn(() => {
-      throw expectedError;
-    });
-    const { getCachedPluginModuleLoader } = await importPluginModuleLoader(
-      "./plugin-module-loader-cache.js?scope=profile-preparation-error",
-    );
-    const loader = getCachedPluginModuleLoader({
-      cache: new Map(),
-      modulePath: "/repo/extensions/demo/api.ts",
-      importerUrl: import.meta.url,
-      aliasMap: {},
-      tryNative: false,
-      createLoader: asPluginModuleLoaderFactory(createLoader),
-    });
-    let thrown: unknown;
-    try {
-      loader("/repo/extensions/demo/api.ts");
-    } catch (error) {
-      thrown = error;
-    }
-    expect(thrown).toBe(expectedError);
-    expect(createLoader).toHaveBeenCalledOnce();
-    expect(output).toHaveBeenCalledExactlyOnceWith(
-      expect.stringMatching(
-        /^\[plugin-load-profile\] phase=source-transform-prepare plugin=\(core\) elapsedMs=\d+\.\d source=\(module\)$/,
-      ),
-    );
-  });
-
   it("normalizes Windows absolute paths when native loading is disabled", async () => {
     vi.spyOn(process, "platform", "get").mockReturnValue("win32");
     const fromSourceTransformer = vi.fn(() => ({ fromSourceTransform: true }));
